Include moon object, update headlights
Included the moon object, which is scenery and is included in all outdoor locations.  Created Fardrop class which is for these types of objects.  Also created subclasses of the Place class to allow us to classify rooms as outdoors or indoors.

Updated the headlights so we can examine them while they are coming at us.